More work on FSA CDK.
The Env lambda is inside a REST API, so it should be a REST thing. In this case, it's now set up to return a JSON.
Updated HTML to load that JSON first.

This isn't working (failes with an internal server error) but I'm pretty sure that's because APIGW doesn't like Lambda's response? I can't get APIGW's logs for that...

Reorganized the stack & constructs. Constructs should never modify their parent, that leads to chaos.

Still can't get COGNITO_APP_CLIENT_ID decoupled, but BASE_URL is good.